+++ django_backend/chatbot_be/views/scrape.py
